% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConfig.R
\name{generate.margins}
\alias{generate.margins}
\title{Sample valid margins for a list of subspaces (not used anymore)}
\usage{
generate.margins(subspaces, values = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.8, 0.7,
  0.9))
}
\arguments{
\item{subspaces}{A list of generated subspaces.}

\item{values}{A vector of valid values for subspace margins.}
}
\value{
A vector containing margins values for each subspaces
}
\description{
Sample valid margins for a list of subspaces (not used anymore)
}
\examples{
generate.margins(list(c(1,2,3), c(4,5), c(5,6)), values=c(0.1,0.2,0.3,0.4,0.5,0.8,0.7,0.9))

}
\author{
Edouard Fouché, \email{edouard.fouche@kit.edu}
}
