% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConfig.R
\name{generate.stream.config}
\alias{generate.stream.config}
\title{Generate a valid stream configuration.}
\usage{
generate.stream.config(dim = 20, mindim = 2, maxdim = 4, values = c(0.1,
  0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9), nsteps = 10, cycle = 0,
  volatility = 0.1)
}
\arguments{
\item{dim}{Number of dimensions in the stream.}

\item{mindim}{Minimum number of dimensions for each subspaces. Should be >= 2.}

\item{maxdim}{Maximum number of dimensions for each subspaces. Should be >= mindim and <= dim/2.}

\item{values}{A vector of valid values for subspace margins.}

\item{nsteps}{Number of steps for which we generate new subspaces and margin values.}

\item{cycle}{Number of iterations to use to create a cyclic streams. If 0, then the stream has no cycle.}

\item{volatility}{Proportion of subspaces to change at any step. Should be > 0 and <= 1. 1 means that all subspaces and margins at changed a each step}
}
\value{
A stream.config object to be passed to a generate.dynamic.stream or generate.static.stream method
}
\description{
Generate a valid stream configuration.
}
\examples{
stream.config <- generate.stream.config(dim=50, nsteps=10, volatility=0.5)

}
\seealso{
\itemize{
\item \code{\link{generate.dynamic.stream}} : generate a dynamic stream
\item \code{\link{generate.static.stream}}  : generate a static stream
}
}
\author{
Edouard Fouché, \email{edouard.fouche@kit.edu}
}
