% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputStream.R
\name{output.stream}
\alias{output.stream}
\title{Output the Stream and its Description}
\usage{
output.stream(stream, prefix)
}
\arguments{
\item{stream}{An object of class stream.}

\item{prefix}{Prefix to give the output files}
}
\value{
Nothing
}
\description{
Write in a bunch of files the result of the stream generation. Output the data as a .csv file
in a <prefix>_data.txt file. Output the labels in a <prefix>_labels.txt file. Output an
extensive description of the data, outliers and subspaces in a <prefix>_description.txt file
}
\examples{
stream <- generate.dynamic.stream()
output.stream(stream, "example")

}
\author{
Edouard Fouché, \email{edouard.fouche@kit.edu}
}
