% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgenerator.R
\docType{package}
\name{streamgenerator}
\alias{streamgenerator}
\alias{streamgenerator-package}
\alias{streamgenerator-package}
\title{streamgenerator}
\description{
A statistical muldi-dimensional stream generator for benchmarking stream mining algorithms.
}
\seealso{
Currently the most useful functions of this package:
\itemize{
\item \code{\link{generate.dynamic.stream}} : generate a dynamic stream
\item \code{\link{generate.static.stream}}  : generate a static stream
\item \code{\link{generate.stream.config}}  : generate a stream.config file for a dynamic or static stream
}
}
\author{
\strong{Maintainer}: Edouard Fouché \email{edouard.fouche@kit.edu}

}
