% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRow.R
\name{generate.multiple.rows}
\alias{generate.multiple.rows}
\title{Helper function that generate multiple rows with the same characteristics
Useful to generate a static stream}
\usage{
generate.multiple.rows(n, dim, subspaces, margins, prop)
}
\arguments{
\item{n}{Number of rows to generate.}

\item{dim}{Number of dimension of the vector generate.}

\item{subspaces}{List of subspaces that contain a hidden space (i.e they show some dependency).}

\item{margins}{List of margins that correspond to each subspace.}

\item{prop}{Probability of a point belonging to the hidden space of a subspace to become an outlier.}
}
\value{
A list with 2 elements where \code{data} is a data.frame object containing the \code{n} vectors and \code{labels} containing \code{n} corresponding labels.
}
\description{
Helper function that generate multiple rows with the same characteristics
Useful to generate a static stream
}
\author{
Edouard Fouché, \email{edouard.fouche@kit.edu}
}
